+++ packages/contracts/test/plugin/counter-plugin-setup.ts
@@ -40,7 +40,7 @@ describe('CounterPluginSetup(Example)', function () {
   let address1: string;
   let address2: string;
 
-  async function decodeEvent(tx: any, eventName: string) {
+  async function findEvent(tx: any, eventName: string) {
     const {events} = await tx.wait();
     const event = events.find(
       ({event}: {event: any}) => event === eventName